## Common: Added the AppConfig to the core, it is probably better to find another way to configure the engine. Cause the Engine could not be compiled by itself, which means it needed a AppConfig, from a project using the Engine, with this done the engine can be easily built as a library.
Added a #pragma once to almos all files, in desperation for improving compile times. (Need work)
## DT3Core/3rdParty:
Modified the script to be able to output libraries in an specific location, having in mind multiple os libraries.
Also added SDL2 compressed sdl2.tar.gz which needs to be extracted in the folder.

## Linux:
Fixed Stream related problem with some operators.
Mostly CMakefiles.

Also added some code in Plaform for sdl2. which needs to be organized.